Refactor and restructure file directory
