+++ legacy/clustering/plane_fitter.py
@@ -1,0 +1,66 @@
+# The MIT License (MIT)
+
+# Copyright (c) 2013 Falcon Dai
+
+# Permission is hereby granted, free of charge, to any person obtaining a copy of
+# this software and associated documentation files (the "Software"), to deal in
+# the Software without restriction, including without limitation the rights to
+# use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+# the Software, and to permit persons to whom the Software is furnished to do so,
+# subject to the following conditions:
+
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+# FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+# COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+# IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+# CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+# https://github.com/falcondai/py-ransac
+
+import numpy as np
+from ransac import *
+
+def augment(xyzs):
+    axyz = np.ones((len(xyzs), 4))
+    axyz[:, :3] = xyzs
+    return axyz
+
+def estimate(xyzs):
+    axyz = augment(xyzs[:3])
+    return np.linalg.svd(axyz)[-1][-1, :]
+
+def is_inlier(coeffs, xyz, threshold):
+    return np.abs(coeffs.dot(augment([xyz]).T)) < threshold
+
+if __name__ == '__main__':
+    import matplotlib
+    import matplotlib.pyplot as plt
+    from mpl_toolkits import mplot3d
+    fig = plt.figure()
+    ax = mplot3d.Axes3D(fig)
+
+    def plot_plane(a, b, c, d):
+        xx, yy = np.mgrid[:10, :10]
+        return xx, yy, (-d - a * xx - b * yy) / c
+
+    n = 100
+    max_iterations = 100
+    goal_inliers = n * 0.3
+
+    # test data
+    xyzs = np.random.random((n, 3)) * 10
+    xyzs[:50, 2:] = xyzs[:50, :1]
+
+    ax.scatter3D(xyzs.T[0], xyzs.T[1], xyzs.T[2])
+
+    # RANSAC
+    m, b = run_ransac(xyzs, estimate, lambda x, y: is_inlier(x, y, 0.01), 3, goal_inliers, max_iterations)
+    a, b, c, d = m
+    xx, yy, zz = plot_plane(a, b, c, d)
+    ax.plot_surface(xx, yy, zz, color=(0, 1, 0, 0.5))
+
+    plt.show()
