--- legacy/ransac.py
@@ -1,50 +1,0 @@
-# The MIT License (MIT)
-
-# Copyright (c) 2013 Falcon Dai
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy of
-# this software and associated documentation files (the "Software"), to deal in
-# the Software without restriction, including without limitation the rights to
-# use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
-# the Software, and to permit persons to whom the Software is furnished to do so,
-# subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
-# FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
-# COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
-# IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
-# CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-# https://github.com/falcondai/py-ransac
-
-import random
-
-def run_ransac(data, estimate, is_inlier, sample_size, goal_inliers, max_iterations, stop_at_goal=True, random_seed=None):
-    best_ic = 0
-    best_model = None
-    random.seed(random_seed)
-    # random.sample cannot deal with "data" being a numpy array
-    data = list(data)
-    for i in range(max_iterations):
-        s = random.sample(data, int(sample_size))
-        m = estimate(s)
-        ic = 0
-        for j in range(len(data)):
-            if is_inlier(m, data[j]):
-                ic += 1
-
-        #print(s)
-        #print('estimate:', m,)
-        #print('# inliers:', ic)
-
-        if ic > best_ic:
-            best_ic = ic
-            best_model = m
-            if ic >= goal_inliers and stop_at_goal:
-                break
-    print('took iterations:', i+1, 'best model:', best_model, 'explains:', best_ic)
-    return best_model, best_ic
