+++ peak_picker.py
@@ -3,8 +3,7 @@ import argparse
 import pyopenms as ms
 import numpy as np
 
-# Try window_size=0.02
-def peak_pick(exp, min_req=3, window_size=float('Inf'), small_peak=0.1, strict=True,
+def peak_pick(exp, min_req=1, window_size=float('Inf'), small_peak=0.1, strict=True,
               sequential=True):
     """A custom peak picker for use with im_binning, since PeakPickerHiRes always
     destroys the data. The idea is to get rid of satellite peaks so that matching
@@ -13,19 +12,21 @@ def peak_pick(exp, min_req=3, window_size=float('Inf'), small_peak=0.1, strict=T
     Algorithm:
     1. Sort peaks by increasing m/z.
     2. Create a boolean array with a False for every peak.
-    3. For each peak, go left and right (within a window) until the current peak is less
-        than x% of the most intense local peak.
-    4. Mark all of the peaks as True.
-    5. Create a new peak with its intensity being the sum of the intensities of the newly
-        marked peaks and its m/z being a weighted average.
+    3. Walk from left to right and find a local maximum.
+    4. For each local maximum, go left and right (within a window) until the current peak
+        is less than x% of the local maximum.
+    5. Mark all of the peaks in the peak set as True.
+    6. Create a new peak with intensity being the sum of the intensities of the peaks in
+        the peak set and m/z being an intensity-weighted average.
 
     Alternate algorithm (for sequential=False):
     1. Sort peaks by increasing m/z.
     2. Create a lookup table, storing each peak's intensity and index in the m/z-sorted
         spectrum.
     3. Sort the lookup table by the peaks' intensities, descending.
-    4. Continue the normal algorithm, but instead of iterating through peaks in the order
-        of the m/z-sorted spectrum, go in the order of the intensity-sorted lookup table.
+    4. Continue with the normal algorithm, but instead of iterating through peaks in the
+        order of the m/z-sorted spectrum, go in the order of the intensity-sorted lookup
+        table.
 
     Args:
         exp (MSExperiment): The OpenMS experiment to be peak picked.
@@ -37,9 +38,8 @@ def peak_pick(exp, min_req=3, window_size=float('Inf'), small_peak=0.1, strict=T
             than the initial peak.
         strict (bool): If true, peaks must be non-increasing from the initial peak.
             Otherwise, a single peak is allowed to break this rule.
-        sequential (bool): If true, peaks are iterated through left-to-right in the
-            sorted m/z spectrum. Otherwise, they are iterated through by decreasing
-            intensity.
+        sequential (bool): If true, run the algorithm normally. Otherwise, use the
+            alternate algorithm.
 
     Returns:
         MSExperiment: An OpenMS experiment corresponding to the input, but peak-picked.
@@ -58,6 +58,7 @@ def peak_pick(exp, min_req=3, window_size=float('Inf'), small_peak=0.1, strict=T
         peak_idx = []
 
         if not sequential:
+            # Create the lookup table
             for i in range(num_peaks):
                 peak_idx.append([spec[i].getIntensity(), i])
 
@@ -69,6 +70,7 @@ def peak_pick(exp, min_req=3, window_size=float('Inf'), small_peak=0.1, strict=T
 
         picked = [False] * num_peaks
 
+        # Begin peak picking
         for idx in range(num_peaks):
             i = idx if sequential else peak_idx[idx][1]
             if picked[i]:
@@ -142,6 +144,7 @@ def peak_pick(exp, min_req=3, window_size=float('Inf'), small_peak=0.1, strict=T
             total_position = 0
             for j in range(low_bound, high_bound + 1):
                 picked[j] = True
+                # Calculate the intensity-weighted m/z position
                 total_position += spec[j].getPos() * (spec[j].getIntensity() /
                                                       total_intensity)
 
@@ -163,9 +166,8 @@ if __name__ == '__main__':
 
     exp = ms.MSExperiment()
     print('Loading mzML input file....................', end='', flush=True)
-    ms.MzMLFile().load(args.input + '.mzML', exp)
+    ms.MzMLFile().load(args.input, exp)
     print('Done')
 
-    new_exp = peak_pick(exp, min_req=3, window_size=float('Inf'), small_peak=0.1,
-                        strict=True, sequential=True)
-    ms.MzMLFile().store(args.output + '.mzML', new_exp)
+    new_exp = peak_pick(exp)
+    ms.MzMLFile().store(args.output, new_exp)
