+++ utils/legacy/tsv_to_csv.py
@@ -3,6 +3,6 @@ import csv
 
 tabin = csv.reader(sys.stdin, dialect=csv.excel_tab)
 commaout = csv.writer(sys.stdout, dialect=csv.excel)
+
 for row in tabin:
     commaout.writerow(row)
-
